requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# OpenAI integration
openai>=1.6.0

# Google Cloud integration  
google-cloud-texttospeech>=2.16.0
google-auth>=2.24.0

# Audio processing
pydub>=0.25.1
librosa>=0.10.1

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0  # PostgreSQL async driver

# HTTP client
aiohttp>=3.9.0
httpx>=0.25.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0

# Video processing (for YouTube uploads)
opencv-python-headless>=4.8.0
pillow>=10.1.0

# Configuration and YAML
pyyaml>=6.0.1
toml>=0.10.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# Development tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Monitoring and logging
structlog>=23.2.0
rich>=13.6.0

# Task queue (for batch processing)
celery[redis]>=5.3.0
redis>=5.0.0

# File handling
pathlib2>=2.3.7

# CLI tools
click>=8.1.0
typer>=0.9.0

# Async utilities
anyio>=4.0.0

# Security
cryptography>=41.0.0

# Production WSGI/ASGI
gunicorn>=21.2.0

# Environment management
python-decouple>=3.8